fix(Details): remove Details.Item
