+++ packages/css/src/details.css
@@ -1,4 +1,29 @@
-.ds-details-group {
+/* .ds-details-group {
+  &[data-border] > * {
+    border: var(--dsc-details-border);
+
+    &:first-child,
+    &:first-child > :is(summary, u-summary) {
+      border-top-left-radius: var(--dsc-details-border-radius);
+      border-top-right-radius: var(--dsc-details-border-radius);
+    }
+
+    &:last-child,
+    &:last-child:not([open]) > :is(summary, u-summary) {
+      border-bottom-left-radius: var(--dsc-details-border-radius);
+      border-bottom-right-radius: var(--dsc-details-border-radius);
+    }
+  }
+} */
+
+.ds-details {
+  --dsc-details-border: 1px solid var(--dsc-details-border-color);
+  --dsc-details-chevron-gap: var(--ds-spacing-2);
+  --dsc-details-chevron-size: var(--ds-spacing-6);
+  --dsc-details-chevron-url: url("data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 24 24'%3E%3Cpath d='M5.97 9.47a.75.75 0 0 1 1.06 0L12 14.44l4.97-4.97a.75.75 0 1 1 1.06 1.06l-5.5 5.5a.75.75 0 0 1-1.06 0l-5.5-5.5a.75.75 0 0 1 0-1.06'/%3E%3C/svg%3E");
+  --dsc-details-padding: var(--ds-spacing-2) var(--ds-spacing-4);
+  --dsc-details-size: var(--ds-sizing-14);
+
   /* default color: neutral */
   --dsc-details-background: var(--ds-color-neutral-background-default);
   --dsc-details-heading-background--hover: var(--ds-color-neutral-surface-default);
@@ -28,33 +53,6 @@
     --dsc-details-heading-background: var(--ds-color-neutral-background-subtle);
   }
 
-  --dsc-details-border: 1px solid var(--dsc-details-border-color);
-  --dsc-details-border-radius: var(--ds-border-radius-md);
-  --dsc-details-chevron-gap: var(--ds-spacing-2);
-  --dsc-details-chevron-size: var(--ds-spacing-6);
-  --dsc-details-chevron-url: url("data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 24 24'%3E%3Cpath d='M5.97 9.47a.75.75 0 0 1 1.06 0L12 14.44l4.97-4.97a.75.75 0 1 1 1.06 1.06l-5.5 5.5a.75.75 0 0 1-1.06 0l-5.5-5.5a.75.75 0 0 1 0-1.06'/%3E%3C/svg%3E");
-
-  --dsc-details-padding: var(--ds-spacing-2) var(--ds-spacing-4);
-  --dsc-details-size: var(--ds-sizing-14);
-
-  &[data-border] > * {
-    border: var(--dsc-details-border);
-
-    &:first-child,
-    &:first-child > :is(summary, u-summary) {
-      border-top-left-radius: var(--dsc-details-border-radius);
-      border-top-right-radius: var(--dsc-details-border-radius);
-    }
-
-    &:last-child,
-    &:last-child:not([open]) > :is(summary, u-summary) {
-      border-bottom-left-radius: var(--dsc-details-border-radius);
-      border-bottom-right-radius: var(--dsc-details-border-radius);
-    }
-  }
-}
-
-.ds-details__item {
   background: var(--dsc-details-background);
   border-block: var(--dsc-details-border);
   box-sizing: border-box;
@@ -76,6 +74,7 @@
 
     &:focus-visible {
       position: relative; /* Ensure foucs outline renders on top */
+      z-index: 1;
     }
 
     &::before {
@@ -90,7 +89,8 @@
   }
 
   & + & {
-    border-top: 0; /* Skip border-top when .details__item is followed by .details__item */
+    border-top: 0; /* Skip border-top when .ds-details is followed by .ds-details */
+    margin-top: 0; /* Prevent gap when places as child of .ds-card */
   }
 
   & > :not(summary, u-summary) {
