+++ packages/react/src/components/Details/Details.mdx
@@ -114,10 +114,6 @@ Ikke plasser andre interaktive elementer inn i `Details`-heading, da hele raden 
 
 ## Andre props
 
-### Details.Item
-
-<ArgTypes of={Details.Item} />
-
 ### Details.Summary
 
 <ArgTypes of={Details.Summary} />
