+++ packages/react/src/components/Details/Details.tsx
@@ -1,42 +1,72 @@
+import { useMergeRefs } from '@floating-ui/react';
 import cl from 'clsx/lite';
 import type { HTMLAttributes, ReactNode } from 'react';
-import { forwardRef } from 'react';
-import type { Color } from '../../colors';
-import type { DefaultProps } from '../../types';
-import type { MergeRight } from '../../utilities';
+import { forwardRef, useEffect, useRef } from 'react';
+import '@u-elements/u-details';
 
-export type DetailsProps = MergeRight<
-  DefaultProps & HTMLAttributes<HTMLDivElement>,
-  {
-    /**
-     * Details background color.
-     * @default neutral
-     */
-    'data-color'?: 'subtle' | Color;
-    /**
-     * Show border
-     * @default false
-     **/
-    border?: boolean;
-    /** Instances of `Details.Item` */
-    children: ReactNode;
-  }
->;
+export type DetailsProps = {
+  /**
+   * Controls open-state.
+   *
+   * Using this removes automatic control of open-state
+   *
+   * @default undefined
+   */
+  open?: boolean;
+  /**
+   * Defaults the details to open if not controlled
+   * @default false
+   */
+  defaultOpen?: boolean;
+  /** Callback function when Details toggles due to click on summary or find in page-search */
+  onToggle?: (event: Event) => void;
+  /** Content should be one `<Details.Summary>` and `<Details.Content>` */
+  children?: ReactNode;
+} & Omit<HTMLAttributes<HTMLDetailsElement>, 'onToggle'> &
+  (
+    | { open: boolean; onToggle: (event: Event) => void }
+    | { open?: never; onToggle?: (event: Event) => void }
+  );
 
 /**
- * Details component, contains `Details.Item` components.
+ * Details component, contains `Details.Summary` and `Details.Content` components.
+ * @example
+ * <Details>
+ *  <DetailsSummary>Header</DetailsSummary>
+ *  <DetailsContent>Content</DetailsContent>
+ * </Details>
  */
-export const Details = forwardRef<HTMLDivElement, DetailsProps>(
+export const Details = forwardRef<HTMLDetailsElement, DetailsProps>(
   function Details(
-    { border = false, 'data-color': color = 'neutral', className, ...rest },
+    { className, open, defaultOpen = false, onToggle, ...rest },
     ref,
   ) {
+    const detailsRef = useRef<HTMLDetailsElement>(null);
+    const initialOpen = useRef(defaultOpen); // Allow rendering defaultOpen on server, but only render once on client
+    const mergedRefs = useMergeRefs([detailsRef, ref]);
+    const onToggleRef = useRef(onToggle); // Using ref to enable access inside useEffect without re-binding event listeners
+    const openRef = useRef(open);
+    onToggleRef.current = onToggle;
+    openRef.current = open;
+
+    // Provide onToggle event and controlled state
+    useEffect(() => {
+      const details = detailsRef.current;
+      const handleToggle = (event: Event) => {
+        if (!details || details?.open === openRef.current) return;
+        onToggleRef.current?.(event);
+        if (openRef.current !== undefined) details.open = openRef.current; // Don't update DOM unless controlled state changes
+      };
+
+      details?.addEventListener('toggle', handleToggle, true);
+      return () => details?.removeEventListener('toggle', handleToggle, true);
+    }, []);
+
     return (
-      <div
-        className={cl('ds-details-group', className)}
-        data-border={border || undefined} /* Fallback to  */
-        data-color={color}
-        ref={ref}
+      <u-details
+        class={cl('ds-details', className)} // Using class since React does not translate className on custom elements
+        open={(open ?? initialOpen.current) || undefined} // Fallback to undefined to prevent rendering open="false"
+        ref={mergedRefs}
         {...rest}
       />
     );
